tests/wpt/tests/fedcm/fedcm-button-and-other-account/fedcm-passive-after-active.tentative.https.html

<!DOCTYPE html>
<title>Federated Credential Management API Passive Mode After Active Mode.</title>
<link rel="help" href="https://fedidcg.github.io/FedCM">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<script type="module">
import {request_options_with_mediation_required,
        fedcm_test,
        fedcm_get_and_select_first_account} from '../support/fedcm-helper.sub.js';

fedcm_test(async t => {
  let passive_test_options = request_options_with_mediation_required();
  let active_test_options = request_options_with_mediation_required("manifest_with_rp_mode.json");
  active_test_options.identity.mode = "active";

  return test_driver.bless('initiate FedCM request', async function() {
      let first_cred = await fedcm_get_and_select_first_account(t, active_test_options);
      assert_equals(first_cred.token, "mode=active");
      let second_cred = await fedcm_get_and_select_first_account(t, passive_test_options);
      assert_equals(second_cred.token, "token");
  });
}, "Test that the passive mode can succeed after active mode.");

</script>
